Refactor logic for a making a single player game
